+++ src/main/java/net/geforcemods/securitycraft/items/SentryRemoteAccessToolItem.java
@@ -41,8 +41,10 @@ public class SentryRemoteAccessToolItem extends Item {
 	public ActionResult<ItemStack> use(World level, PlayerEntity player, Hand hand) {
 		ItemStack stack = player.getItemInHand(hand);
 
-		if (!level.isClientSide)
-			SecurityCraft.channel.send(PacketDistributor.PLAYER.with(() -> (ServerPlayerEntity) player), new OpenScreen(DataType.SENTRY_REMOTE_ACCESS_TOOL));
+		if (!level.isClientSide) {
+			updateTagWithNames(stack, level);
+			SecurityCraft.channel.send(PacketDistributor.PLAYER.with(() -> (ServerPlayerEntity) player), new OpenScreen(DataType.SENTRY_REMOTE_ACCESS_TOOL, stack.getTag()));
+		}
 
 		return ActionResult.consume(stack);
 	}
@@ -53,16 +55,16 @@ public class SentryRemoteAccessToolItem extends Item {
 		BlockPos pos = ctx.getClickedPos();
 		PlayerEntity player = ctx.getPlayer();
 		List<Sentry> sentries = level.getEntitiesOfClass(Sentry.class, new AxisAlignedBB(pos));
+		ItemStack stack = ctx.getItemInHand();
 
 		if (!sentries.isEmpty()) {
 			Sentry sentry = sentries.get(0);
 			BlockPos sentryPos = sentry.blockPosition();
-			ItemStack stack = ctx.getItemInHand();
 
 			if (!isSentryAdded(stack, sentryPos)) {
-				int availSlot = getNextAvaliableSlot(stack);
+				int nextAvailableSlot = getNextAvaliableSlot(stack);
 
-				if (availSlot == 0) {
+				if (nextAvailableSlot == 0) {
 					PlayerUtils.sendMessageToPlayer(player, Utils.localize(SCContent.SENTRY_REMOTE_ACCESS_TOOL.get().getDescriptionId()), Utils.localize("messages.securitycraft:srat.noSlots"), TextFormatting.RED);
 					return ActionResultType.FAIL;
 				}
@@ -72,10 +74,10 @@ public class SentryRemoteAccessToolItem extends Item {
 					return ActionResultType.FAIL;
 				}
 
-				if (stack.getTag() == null)
-					stack.setTag(new CompoundNBT());
+				stack.getOrCreateTag().putIntArray("sentry" + nextAvailableSlot, BlockUtils.posToIntArray(sentryPos));
 
-				stack.getTag().putIntArray(("sentry" + availSlot), BlockUtils.posToIntArray(sentryPos));
+				if (sentry.hasCustomName())
+					stack.getTag().putString("sentry" + nextAvailableSlot + "_name", sentry.getCustomName().getString());
 
 				if (!level.isClientSide)
 					SecurityCraft.channel.send(PacketDistributor.PLAYER.with(() -> (ServerPlayerEntity) player), new UpdateNBTTagOnClient(stack));
@@ -89,8 +91,10 @@ public class SentryRemoteAccessToolItem extends Item {
 
 			return ActionResultType.SUCCESS;
 		}
-		else if (!level.isClientSide)
-			SecurityCraft.channel.send(PacketDistributor.PLAYER.with(() -> (ServerPlayerEntity) player), new OpenScreen(DataType.SENTRY_REMOTE_ACCESS_TOOL));
+		else if (!level.isClientSide) {
+			updateTagWithNames(stack, level);
+			SecurityCraft.channel.send(PacketDistributor.PLAYER.with(() -> (ServerPlayerEntity) player), new OpenScreen(DataType.SENTRY_REMOTE_ACCESS_TOOL, stack.getTag()));
+		}
 
 		return ActionResultType.SUCCESS;
 	}
@@ -109,19 +113,58 @@ public class SentryRemoteAccessToolItem extends Item {
 					tooltip.add(new StringTextComponent(TextFormatting.GRAY + "---"));
 				else {
 					BlockPos pos = new BlockPos(coords[0], coords[1], coords[2]);
-					List<Sentry> sentries = Minecraft.getInstance().player.level.getEntitiesOfClass(Sentry.class, new AxisAlignedBB(pos));
-					String nameToShow;
+					String nameKey = "sentry" + i + "_name";
+					String nameToShow = null;
 
-					if (!sentries.isEmpty() && sentries.get(0).hasCustomName())
-						nameToShow = sentries.get(0).getCustomName().getString();
-					else
-						nameToShow = Utils.localize("tooltip.securitycraft:sentry").getString() + " " + i;
+					if (stack.getTag().contains(nameKey))
+						nameToShow = stack.getTag().getString(nameKey);
+					else {
+						List<Sentry> sentries = Minecraft.getInstance().player.level.getEntitiesOfClass(Sentry.class, new AxisAlignedBB(pos));
+
+						if (!sentries.isEmpty() && sentries.get(0).hasCustomName())
+							nameToShow = sentries.get(0).getCustomName().getString();
+						else
+							nameToShow = Utils.localize("tooltip.securitycraft:sentry").getString() + " " + i;
+					}
 
 					tooltip.add(new StringTextComponent(TextFormatting.GRAY + nameToShow + ": " + Utils.getFormattedCoordinates(pos).getString()));
 				}
 			}
 			else
 				tooltip.add(new StringTextComponent(TextFormatting.GRAY + "---"));
+		}
+	}
+
+	private void updateTagWithNames(ItemStack stack, World level) {
+		if (!stack.hasTag())
+			return;
+
+		CompoundNBT tag = stack.getTag();
+
+		for (int i = 1; i <= 12; i++) {
+			int[] coords = tag.getIntArray("sentry" + i);
+			String nameKey = "sentry" + i + "_name";
+
+			if (coords.length == 3) {
+				BlockPos sentryPos = new BlockPos(coords[0], coords[1], coords[2]);
+
+				if (level.isLoaded(sentryPos)) {
+					List<Sentry> sentries = level.getEntitiesOfClass(Sentry.class, new AxisAlignedBB(sentryPos));
+
+					if (!sentries.isEmpty()) {
+						Sentry sentry = sentries.get(0);
+
+						if (sentry.hasCustomName()) {
+							tag.putString(nameKey, sentry.getCustomName().getString());
+							continue;
+						}
+					}
+				}
+				else
+					continue;
+			}
+
+			tag.remove(nameKey);
 		}
 	}